requirements.txt
fastapi==0.115.0 
pydantic==2.9.2 
langchain==0.3.0 
langchain-community==0.3.0 
langchain-openai==0.2.0 
python-dotenv==1.0.1 
uvicorn==0.30.6 
serpapi==0.1.5 
motor==3.6.0 
pymongo==4.9
python-jose==3.3.0 
passlib[bcrypt]==1.7.4
google-search-results
pydantic[email]